requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.28.0
pandas>=2.0.0
pyarrow>=14.0.0
requests>=2.31.0
python-dotenv>=1.0.0
httpx>=0.27.0
pytest>=8.0.0
